feat(middleware): apply authMiddleware to friends routes
