+++ server/middleware/auth.js
@@ -1,0 +1,22 @@
+const supabase = require('../utils/supabaseclient');
+
+const authMiddleware = async (req, res, next) => {
+    const authHeader = req.headers.authorization;
+
+    if (!authHeader || !authHeader.startsWith('Bearer ')) {
+        return res.status(401).json({ error: 'Missing or malformed Authorization header' });
+    }
+
+    const token = authHeader.split(' ')[1];
+    const { data: authData, error: authError } = await supabase.auth.getUser(token);
+    const user = authData?.user;
+
+    if (authError || !user) {
+        return res.status(401).json({ error: 'Invalid or expired token' });
+    }
+
+    req.user = user;
+    next();
+};
+
+module.exports = authMiddleware;
