+++ server/routes/friends.js
@@ -9,23 +9,11 @@ const supabase = createClient(
     process.env.EXPO_PUBLIC_SUPABASE_SERVICE_ROLE_KEY
 );
 
+const authMiddleware = require('../middleware/auth');
 
-router.get('/', async (req, res) => {
-    const authHeader = req.headers.authorization;
 
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-        return res.status(401).json({ error: 'Missing or malformed Authorization header' });
-    }
-  
-    const token = authHeader.split(' ')[1];
-    const { data: authData, error: authError } = await supabase.auth.getUser(token);
-    const user = authData?.user;
-  
-    if (authError || !user) {
-        return res.status(401).json({ error: 'Invalid or expired token' });
-    }
-  
-    const user_id = user.id;
+router.get('/', authMiddleware, async (req, res) => {
+    const user_id = req.user.id;
   
     const { data: friendships, error: friendsError } = await supabase
         .from('friendships')
@@ -59,16 +47,9 @@ router.get('/', async (req, res) => {
 });
 
 // API endpoint for searching users
-router.get('/search', async (req, res) => {
-    const authHeader = req.headers.authorization;
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-        return res.status(401).json({ error: 'Missing or malformed Authorization header' });
-    }
-  
-    const token = authHeader.split(' ')[1];
-    const { data: authData, error: authError } = await supabase.auth.getUser(token);
-    const user = authData?.user;
+router.get('/search', authMiddleware, async (req, res) => {
     const { username } = req.query;
+    const user = req.user;
   
     if (!username || !user) {
         return res.status(400).json({ error: 'Missing username or unauthorized' });
@@ -146,22 +127,8 @@ router.post('/send-request', async (req, res) => {
 });
 
 // API endpoint to get requests
-router.get('/requests', async (req, res) => {
-    const authHeader = req.headers.authorization;
-
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-        return res.status(401).json({ error: 'Missing or malformed Authorization header' });
-    }
-  
-    const token = authHeader.split(' ')[1];
-    const { data, error } = await supabase.auth.getUser(token);
-    const user = data?.user;
-  
-    if (error || !user) {
-        return res.status(401).json({ error: 'Invalid or expired token' });
-    }
-  
-    const user_id = user.id;
+router.get('/requests', authMiddleware, async (req, res) => {
+    const user_id = req.user.id;
   
     const { data: friends, error: friendError } = await supabase
         .from('friendships')
@@ -229,22 +196,9 @@ router.patch('/reject-request', async (req, res) => {
     res.status(200).json({ message: 'Friend request rejected successfully' });
 });
 
-router.delete("/remove/:targetUserId", async (req, res) => {
-    const authHeader = req.headers.authorization;
-
-    if (!authHeader || !authHeader.startsWith("Bearer ")) {
-        return res.status(401).json({ error: "Missing or malformed Authorization header" });
-    }
-  
-    const token = authHeader.split(" ")[1];
-    const { data: authData, error: authError } = await supabase.auth.getUser(token);
-    const user = authData?.user;
-  
-    if (authError || !user) {
-        return res.status(401).json({ error: "Invalid or expired token" });
-    }
-  
-    const userId = user.id;
+// API endpoint to remove friendship
+router.delete("/remove/:targetUserId", authMiddleware, async (req, res) => {
+    const userId = req.user.id;
     const targetUserId = req.params.targetUserId;
 
     try {
